refactor: rename filter to filterscard
