+++ frontend/svelte/src/lib/components/FiltersCard.svelte
